requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.4.0
structlog>=23.2.0
python-json-logger>=2.0.7

# Browser automation
# browser-use>=0.1.40  # Package not available on PyPI, install separately if needed
playwright>=1.40.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0

# Web scraping
crawl4ai>=0.5.0.post4
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Document processing
python-docx>=1.0.0
docxtpl>=0.16.7
markdown>=3.4.0
PyPDF2>=3.0.0
pdf2image>=1.16.0

# AI/LLM integration
transformers>=4.35.0
torch>=2.1.0
openai>=1.3.0
# spacy>=3.7.0  # Not compatible with Python 3.13 yet
scikit-learn>=1.3.0
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz  # Requires spacy
fuzzywuzzy>=0.18.0
python-dateutil>=2.8.2

# Database and ORM
SQLAlchemy>=2.0.0
alembic>=1.13.0
tenacity>=8.0.0
psycopg2-binary>=2.9.0  # Optional PostgreSQL support
# mysqlclient>=2.2.0      # Optional MySQL support - difficult to install on Windows

# Utility libraries
tqdm>=4.66.0
# asyncio>=3.4.3  # Built-in Python module, not needed in requirements
aiohttp>=3.8.5
httpx>=0.24.1
tenacity>=8.2.0
cachetools>=5.3.0
click>=8.1.0
colorama>=0.4.6
rich>=13.7.0
faiss-cpu
sentence-transformers

# Date/time handling
pytz>=2023.3

# Type annotations
typing-extensions>=4.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Logging and Monitoring
prometheus-client>=0.19.0
textblob

# Google Generative AI
google-generativeai>=0.3.0  # For Gemini API integration
azure-ai-inference>=1.0.0b7

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.36.0